+++ Water_Features/ChangeLog.md
@@ -1,5 +1,5 @@
 ﻿## Patch V1.1.11
-* Added a setting to force water simulation in game to be at least 1x.
+* Added a setting to attempt to force water simulation in game to be at least 1x.
 * Seasonal streams is by default off.
 * Seasonal Streams and Waves and Tides settings are only available in game or in the editor. Whether those options were last enabled or not in a save file determines whether the settings are enabled or disabled after loading. Valid water sources are required or it will default to disabled.
-* Reorganized settings tabs with Water Tool tab as initial tab.
+* Reorganized settings including making Water Tool tab as initial tab.